Reorganize adapters code into modular structure
Split large adapters files into logical modules:

- core.py: Main adapt_node/adapt_tree functions
- utils.py: Error handling decorator and utilities
- extraction/: Advanced extraction functionality
  - engine.py: Main extract_attribute orchestration
  - path_parser.py: Path expression parsing
  - path_evaluator.py: Path evaluation engine
  - transforms.py: Transformation functions
  - filters.py: Collection filtering

Benefits:
- Better separation of concerns
- Easier to understand and maintain
- Individual components can be tested in isolation
- More extensible for new transformations/filters

API compatibility maintained - all public functions available from treeviz.adapters
