+++ python/tests/treeviz/advanced_extraction/test_advanced_extraction.py
@@ -6,7 +6,7 @@ complex path expressions, transformations, filtering, and fallback chains.
 """
 
 import pytest
-from treeviz.adapters.advanced_extraction import (
+from treeviz.adapters.extraction import (
     extract_by_path,
     apply_transformation,
     filter_collection,
